requirements.txt
fastapi
uvicorn
transformers
peft
torch
python-dotenv
unsloth